% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{n}{the number of between-sample random numbers}

\item{a}{the first parameter of Beta distribution}

\item{b}{the second parameter of Beta distribution}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
set.seed(1230)
NumericMatrix gibbsC(1000,1000,1,1)
}
}
