% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21059R.R
\name{rej.perc}
\alias{rej.perc}
\title{rejpercent}
\usage{
rej.perc(m, sigma)
}
\arguments{
\item{m}{the length of chain}

\item{sigma}{the parameter of Rayleigh distribution}
}
\value{
the rejection percent of the final chain
}
\description{
use the Metropolis-Hastings sampler to generate a sample from a Rayleigh distribution,and then evaluate the rejection percent of the candidate point
}
\examples{
\dontrun{
m<-1000
sigma<-5
rej.perc(m,sigma)
}
}
