\name{StatComp21059-package}
\alias{StatComp21059-package}
\alias{StatComp21059}
\docType{package}
\title{
  Evaluation of variance reduction and generation of chain
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Tong Zhi <zhit@mail.ustc.edu.cn>

Maintainer: Tong Zhi <zhit@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
