% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp21059R.R
\name{var.redu}
\alias{var.redu}
\title{varreduction}
\usage{
var.redu(m, x, R)
}
\arguments{
\item{m}{times of repetition}

\item{x}{upper limit of Monte Carlo Integration}

\item{R}{the sample size}
}
\value{
the reduction of the variance through using the antithetic variables in Monte Carlo Integration
}
\description{
Implement a function to generate samples from standard normal distribution,and then use antithetic variables.In the end we will get the percent of variance reduction
}
\examples{
\dontrun{
m<-1000
R<-10000
x<-2
var.redu(m,x,R)
}
}
